requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
jinja2==3.1.6
python-multipart==0.0.20
python-dotenv==1.1.1
playwright==1.54.0
openai==1.99.6
itsdangerous==2.2.0
